+++ %ProgramData%/Microsoft/Windows/UUS/State/_active.uusver
@@ -1,0 +1,1 @@
+1216.2402.13012.0
