+++ app/(auth)/_layout.tsx
@@ -1,0 +1,31 @@
+import { Redirect, Stack } from "expo-router";
+import { StatusBar } from "expo-status-bar";
+
+// import { Loader } from "../../components";
+// import { useGlobalContext } from "../../context/GlobalProvider";
+
+const AuthLayout = () => {
+  //   const { loading, isLogged } = useGlobalContext();
+
+  //   if (!loading && isLogged) return <Redirect href="/home" />;
+
+  return (
+    <>
+      <Stack>
+        <Stack.Screen
+          name="login"
+          options={{
+            headerShown: false,
+          }}
+        />
+
+        <Stack.Screen name="profile" />
+      </Stack>
+
+      {/* <Loader isLoading={loading} />
+      <StatusBar backgroundColor="#161622" style="light" /> */}
+    </>
+  );
+};
+
+export default AuthLayout;
