+++ app/onboardingScreen.tsx
@@ -1,20 +1,14 @@
-import {
-  Dimensions,
-  Image,
-  StatusBar,
-  StyleSheet,
-  Text,
-  View,
-} from "react-native";
+import { Image, StatusBar, StyleSheet, View } from "react-native";
 import React from "react";
 import { OnboardFlow } from "react-native-onboard";
+import { router } from "expo-router";
 
-const OnboardingScreen = ({ navigation }: any) => {
+const OnboardingScreen = () => {
   return (
     <View style={styles.container}>
       <StatusBar barStyle="dark-content" />
       <OnboardFlow
-        onDone={() => navigation.navigate("Login")}
+        onDone={() => router.navigate("/(auth)/login")}
         subtitleStyle={styles.subTitle}
         titleStyle={styles.title}
         pages={[
@@ -22,21 +16,21 @@ const OnboardingScreen = ({ navigation }: any) => {
             title: "Hello User!",
             subtitle: "Welcome To Sarvarath",
             imageUri: Image.resolveAssetSource(
-              require("../assets/images/logo.png"),
+              require("../assets/images/logo.png")
             ).uri,
           },
           {
             title: "Find nearby Buses",
             subtitle: "Remove any hassle and find the bus to your destination",
             imageUri: Image.resolveAssetSource(
-              require("../assets/images/nearbyBus.png"),
+              require("../assets/images/nearbyBus.png")
             ).uri,
           },
           {
             title: "Track Buses",
             subtitle: "A convenient way to track the location of the buses",
             imageUri: Image.resolveAssetSource(
-              require("../assets/images/trackBus.jpg"),
+              require("../assets/images/trackBus.jpg")
             ).uri,
           },
           {
@@ -44,14 +38,14 @@ const OnboardingScreen = ({ navigation }: any) => {
             subtitle:
               "Get detailed information about the stops and routes of any bus",
             imageUri: Image.resolveAssetSource(
-              require("../assets/images/busDetails.jpg"),
+              require("../assets/images/busDetails.jpg")
             ).uri,
           },
           {
             title: "Let's get you started!",
             subtitle: "Login to continue",
             imageUri: Image.resolveAssetSource(
-              require("../assets/images/getStarted.jpg"),
+              require("../assets/images/getStarted.jpg")
             ).uri,
           },
         ]}
