--- screens/index.tsx
@@ -1,7 +1,0 @@
-import HomePage from "./homePage";
-import Profile from "./profile";
-import SearchBuses from "./searchBuses";
-import Login from "./login";
-import MapScreen from "./mapScreen";
-
-export { Profile, HomePage, SearchBuses, Login, MapScreen };
